fix(sdds-acore/uikit-compose): Tooltip clickable area was fixed
